+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This project uses a **major**.**minor**.**micro** versioning scheme, where:
 * The minor version is bumped on minor changes to the `omst` crate, as defined by Rust RFC 1122.
 * The micro version is bumped in all other cases.
 
+# v3.0.0
+
+* [changed] `omst` function now returns an `io::Result`
+* [changed] internal implementations are exposed via `winapi` and `shadow` modules
+
 # v2.0.0
 
 * [added] `omst-be` binary which fully prints errors